# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a golyadmin Server",
        "title": "Swagger API",
        "contact": {},
        "version": "0.0.1"
    },
    "paths": {
...
        "/system/menu_button/menu_button": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MenuButton"
                ],
                "summary": "获取菜单按钮全部列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "方法",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取菜单按钮全部列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MenuButton"
                ],
                "summary": "新增菜单按钮",
                "parameters": [
                    {
                        "description": "名称",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.LyadminMenuButton"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "新增菜单按钮",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MenuButton"
                ],
                "summary": "根据ID删除菜单按钮",
                "parameters": [
                    {
                        "description": "LyadminMenuButton模型",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除LyadminMenuButton",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/menu_button/menu_button/:id": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MenuButton"
                ],
                "summary": "编辑菜单按钮",
                "parameters": [
                    {
                        "description": "名称",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/system.LyadminMenuButton"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "编辑菜单按钮",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/system/menu_button/pagelist": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MenuButton"
                ],
                "summary": "分页获取菜单按钮列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "每页大小",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "方法",
                        "name": "method",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "当前页码",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "分页获取菜单按钮列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.StructResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": true
                                        },
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.Id": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer"
                }
            }
        },
...
        "response.StructResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "system.LyadminMenuButton": {
            "type": "object",
            "properties": {
                "api": {
                    "type": "string"
                },
                "belong_dept": {
                    "description": "数据归属部门",
                    "type": "integer"
                },
                "create_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "create_by": {
                    "description": "创建者",
                    "type": "integer"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "menu_id": {
                    "type": "integer"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "update_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "update_by": {
                    "description": "更新着",
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}